Add repo settings

+++ interface/models.py
@@ -36,31 +36,40 @@ class Repo(models.Model):
     full_name = models.TextField(unique=True)
     webhook_id = models.IntegerField(null=True, blank=True)
     is_private = models.BooleanField(default=True)
+    default_branch = models.TextField(default='master')
 
+    disabled = models.BooleanField(default=False)
     created_at = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
         return self.full_name
+
+    def get_absolute_url(self):
+        return reverse('repo_detail', kwargs={'full_name': self.full_name})
 
     @property
     def clone_url(self):
         return 'https://github.com/{}.git'.format(self.full_name)
 
     def get_head_build(self):
-        return Build.objects.filter(repo=self, ref='master').only('status').first()
+        return Build.objects.filter(repo=self, ref=self.default_branch).only('status').first()
 
     def soft_delete(self):
-        g = get_github(self.user)
-        grepo = g.get_repo(self.full_name)
+        self.disabled = True
+        self.remove_webhook()
 
-        try:
-            assert self.webhook_id
-            hook = grepo.get_hook(self.webhook_id)
-            hook.delete()
-            self.webhook_id = None
-        except (UnknownObjectException, AssertionError):
-            pass
+    def remove_webhook(self):
+        if not settings.DEBUG:
+            g = get_github(self.user)
+            grepo = g.get_repo(self.full_name)
 
+            try:
+                hook = grepo.get_hook(self.webhook_id)
+                hook.delete()
+            except UnknownObjectException:
+                pass
+
+        self.webhook_id = None
         self.save()
 
     def user_is_collaborator(self, user):
@@ -74,22 +83,24 @@ class Repo(models.Model):
         return grepo.has_in_collaborators(guser)
 
     def add_webhook(self, request):
-        g = get_github(request.user)
-        grepo = g.get_repo(self.full_name)
+        if settings.DEBUG:
+            self.webhook_id = 123
+        else:
+            g = get_github(self.user)
+            grepo = g.get_repo(self.full_name)
 
-        hook = grepo.create_hook(
-            'web',
-            {
-                'content_type': 'json',
-                'url': request.build_absolute_uri(reverse('webhook')),
-                'secret': settings.WEBHOOK_SECRET
-            },
-            events=['push'],
-            active=True
-        )
+            hook = grepo.create_hook(
+                'web',
+                {
+                    'content_type': 'json',
+                    'url': request.build_absolute_uri(reverse('webhook')),
+                    'secret': settings.WEBHOOK_SECRET
+                },
+                events=['push'],
+                active=True
+            )
+            self.webhook_id = hook.id
 
-        self.webhook_id = hook.id
-        self.is_private = grepo.private
         self.save()
 
     class Meta:
@@ -154,7 +165,7 @@ class Build(models.Model):
         if self.status != state:
             self.status = state
             self.save()
-        if not settings.DEBUG:
+        if not settings.DEBUG and self.repo.webhook_id:
             self.publish_status(auth, state, message)
 
     def clone(self, auth):
